+++ src/hyperfine/error.rs
@@ -12,6 +12,7 @@ pub enum ParameterScanError {
     TooLarge,
     ZeroStep,
     StepRequired,
+    UnexpectedCommandNameCount(usize, usize),
 }
 
 impl From<num::ParseIntError> for ParameterScanError {
@@ -40,6 +41,13 @@ impl fmt::Display for ParameterScanError {
                  floating point numbers. The step size can be specified \
                  with the '--parameter-step-size' parameter"
             ),
+            ParameterScanError::UnexpectedCommandNameCount(real, expected) => {
+                write!(
+                    f,
+                    "'--command-name' has been specified {} times. It has to appear exactly once, or exactly {} times (number of benchmarks)",
+                    real, expected
+                )
+            }
         }
     }
 }
@@ -51,6 +59,7 @@ pub enum OptionsError<'a> {
     RunsBelowTwo,
     EmptyRunsRange,
     TooManyCommandNames(usize),
+    UnexpectedCommandNameCount(usize, usize),
     NumericParsingError(&'a str, ParseIntError),
 }
 
@@ -61,6 +70,13 @@ impl<'a> fmt::Display for OptionsError<'a> {
             OptionsError::RunsBelowTwo => write!(f, "Number of runs below two"),
             OptionsError::TooManyCommandNames(n) => {
                 write!(f, "Too many --command-name options: expected {} at most", n)
+            }
+            OptionsError::UnexpectedCommandNameCount(real, expected) => {
+                write!(
+                    f,
+                    "'--command-name' has been specified {} times. It has to appear exactly once, or exactly {} times (number of benchmarks)",
+                    real, expected
+                )
             }
             OptionsError::NumericParsingError(cmd, ref err) => write!(
                 f,