+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- `--command-name` can now take parameter names from `--parameter-*` options, see #351 and #391 (@silathdiir)
 - Add command exit code to output if it fails, see #342 (@KaindlJulian)
 - Export command exit code to JSON output, see #371 (@JordiChauzi)
 
@@ -11,6 +12,7 @@
   to fail early in case of, e.g., wrong permissions. See #306 (@s1ck).
 - When `--export-*` options are used, result files are written after each individual
   benchmark command instead of writing after all benchmarks have finished. See #306 (@s1ck).
+- Reduce number of shell startup time measurements from 200 to 50, generally speeding up benchmarks. See #378
 
 ## Bugfixes